Fix cdai handling
