+++ tests/unit/executor_test.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from backd import executor
 
-from tests.conftest import DUMMY_MARKETS_META
+from tests.fixtures import DUMMY_MARKETS_META
 
 
 MAIN_MARKET = "0x1A3B"
@@ -11,8 +11,7 @@ MAIN_MARKET = "0x1A3B"
 
 @patch("backd.protocols.compound.constants.MARKETS", DUMMY_MARKETS_META)
 def test_process_all_events():
-    state = executor.process_all_events(
-        "compound", min_block=120, max_block=125)
+    state = executor.process_all_events("compound", min_block=120, max_block=125)
     market = state.markets.find_by_address(MAIN_MARKET)
     liquidator_user_balance = market.users["0xab31"].balances
     assert liquidator_user_balance.token_balance == 55
